+++ Classes/CreateCoinViewController.m
@@ -838,7 +838,7 @@ static sqlite3 *database = nil;
 	size1.width=205;
 	size1.height=100;
 	appDelegate.coinController.currentButton = (UIButton *)sender;
-	tempID = appDelegate.coinController.currentButton.tag;
+	tempID = (int)appDelegate.coinController.currentButton.tag;
 	
 	UIViewController *choices = [[UIViewController alloc] init];
 	//-----Images Button----
@@ -899,7 +899,7 @@ static sqlite3 *database = nil;
 	size2.width=320;
 	size2.height=484;
 	//appDelegate.viewController.currentButton = (UIButton *)sender;
-	tempID = appDelegate.coinController.currentButton.tag;
+	tempID = (int)appDelegate.coinController.currentButton.tag;
 	NSLog(@"currentButton tag (ROUND1): %d",tempID);
 	
 	
@@ -1075,7 +1075,7 @@ static sqlite3 *database = nil;
 	IPAACAppDelegate *appDelegate = 
 	[[UIApplication sharedApplication] delegate];
 	
-	tempID = appDelegate.coinController.currentButton.tag;
+	tempID = (int)appDelegate.coinController.currentButton.tag;
 	
 	NSArray *paths = NSSearchPathForDirectoriesInDomains(NSDocumentDirectory , NSUserDomainMask, YES);
 	NSString *documentsDir = [paths objectAtIndex:0];
@@ -1124,7 +1124,7 @@ static sqlite3 *database = nil;
 	
 	IPAACAppDelegate *appDelegate = 
 	[[UIApplication sharedApplication] delegate];
-	tempID = appDelegate.coinController.currentButton.tag;
+	tempID = (int)appDelegate.coinController.currentButton.tag;
 	
 	NSString *tempStr = [detailAudio description];
 	NSArray *temp = [tempStr componentsSeparatedByString:@"."];
